requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
numpy==1.20.0
pefile==2019.4.18
Pillow==8.0.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pywin32-ctypes==0.2.0
requests==2.24.0
soupsieve==2.0.1
srcomapi==0.3.1
tqdm==4.51.0
urllib3==1.25.11
youtube-dl==2021.1.8
